Release v2.6.3

+++ post-expirator.php
@@ -4,14 +4,14 @@ Plugin Name: Post Expirator
 Plugin URI: http://wordpress.org/extend/plugins/post-expirator/
 Description: Allows you to add an expiration date (minute) to posts which you can configure to either delete the post, change it to a draft, or update the post categories at expiration time.
 Author: PublishPress
-Version: 2.6.2
+Version: 2.6.3
 Author URI: http://publishpress.com
 Text Domain: post-expirator
 Domain Path: /languages
 */
 
 // Default Values
-define('POSTEXPIRATOR_VERSION', '2.6.2');
+define('POSTEXPIRATOR_VERSION', '2.6.3');
 define('POSTEXPIRATOR_DATEFORMAT', __('l F jS, Y', 'post-expirator'));
 define('POSTEXPIRATOR_TIMEFORMAT', __('g:ia', 'post-expirator'));
 define('POSTEXPIRATOR_FOOTERCONTENTS', __('Post expires at EXPIRATIONTIME on EXPIRATIONDATE', 'post-expirator'));
@@ -677,12 +677,12 @@ function postexpirator_schedule_event($id, $ts, $opts)
 
     $error = wp_schedule_single_event($ts, 'postExpiratorExpire', array($id), true);
     if (POSTEXPIRATOR_DEBUG) {
+
         $debug->save(
             array(
-                'message' => $id . ' -> SCHEDULED at ' . date_i18n(
-                        'r',
-                        $ts
-                    ) . ' ' . '(' . $ts . ') with options ' . print_r($opts, true) . ' ' . (is_wp_error(
+                'message' => $id . ' -> SCHEDULED at ' .
+                    PostExpirator_Util::get_wp_date('r', $ts)
+                    . ' ' . '(' . $ts . ') with options ' . print_r($opts, true) . ' ' . (is_wp_error(
                         $error
                     ) ? $error->get_error_message() : 'no error')
             )
@@ -1448,7 +1448,7 @@ function postexpirator_shortcode($atts)
         $format = $timeformat;
     }
 
-    return date_i18n($format, $expirationdatets + (get_option('gmt_offset') * HOUR_IN_SECONDS));
+    return PostExpirator_Util::get_wp_date($format, $expirationdatets);
 }
 
 add_shortcode('postexpirator', 'postexpirator_shortcode');
@@ -1487,10 +1487,11 @@ function postexpirator_add_footer($text)
         'EXPIRATIONDATE',
         'EXPIRATIONTIME',
     );
+
     $replace = array(
-        get_date_from_gmt(gmdate('Y-m-d H:i:s', $expirationdatets), "$dateformat $timeformat"),
-        get_date_from_gmt(gmdate('Y-m-d H:i:s', $expirationdatets), $dateformat),
-        get_date_from_gmt(gmdate('Y-m-d H:i:s', $expirationdatets), $timeformat),
+        PostExpirator_Util::get_wp_date("$dateformat $timeformat", $expirationdatets),
+        PostExpirator_Util::get_wp_date($dateformat, $expirationdatets),
+        PostExpirator_Util::get_wp_date($timeformat, $expirationdatets)
     );
 
     $add_to_footer = '<p style="' . $expirationdateFooterStyle . '">' . str_replace(
@@ -1519,7 +1520,7 @@ function postexpirator_debug()
         if (! defined('POSTEXPIRATOR_DEBUG')) {
             define('POSTEXPIRATOR_DEBUG', 1);
         }
-        require_once(plugin_dir_path(__FILE__) . 'post-expirator-debug.php'); // Load Class
+        require_once(POSTEXPIRATOR_BASEDIR . '/post-expirator-debug.php'); // Load Class
 
         return new PostExpiratorDebug();
     } else {
@@ -1708,6 +1709,12 @@ function postexpirator_activate()
  */
 function expirationdate_deactivate()
 {
+    $preserveData = (bool)get_option('expirationdatePreserveData', true);
+
+    if ($preserveData) {
+        return;
+    }
+
     global $current_blog;
     delete_option('expirationdateExpiredPostStatus');
     delete_option('expirationdateExpiredPageStatus');
@@ -1727,13 +1734,14 @@ function expirationdate_deactivate()
     delete_option('expirationdateDefaultsPage');
     delete_option('expirationdateDefaultsPost');
     delete_option('expirationdateGutenbergSupport');
+    delete_option('expirationdatePreserveData');
     // what about custom post types? - how to cleanup?
     if (is_multisite()) {
         wp_clear_scheduled_hook('expirationdate_delete_' . $current_blog->blog_id);
     } else {
         wp_clear_scheduled_hook('expirationdate_delete');
     }
-    require_once(plugin_dir_path(__FILE__) . 'post-expirator-debug.php');
+    require_once(POSTEXPIRATOR_BASEDIR . '/post-expirator-debug.php');
     $debug = new PostExpiratorDebug();
     $debug->removeDbTable();
 }
@@ -2058,7 +2066,7 @@ function postexpirator_autoload($class)
     foreach ($namespaces as $namespace) {
         if (substr($class, 0, strlen($namespace)) === $namespace) {
             $class = str_replace('_', '', strstr($class, '_'));
-            $filename = plugin_dir_path(__FILE__) . 'classes/' . sprintf('%s.class.php', $class);
+            $filename = POSTEXPIRATOR_BASEDIR . '/classes/' . sprintf('%s.class.php', $class);
             if (is_readable($filename)) {
                 require_once $filename;
